backend/requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.0.3
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-redis==5.4.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
PyJWT==2.9.0
pymemcache==4.0.0
python-dotenv==1.0.1
python-memcached==1.62
redis==5.2.1
requests==2.32.3
sqlparse==0.5.1
tzlocal==5.2
urllib3==2.2.3
whitenoise==6.7.0
psycopg2-binary==2.9.10

